% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susobasemap-package.R
\docType{package}
\name{susobasemap-package}
\alias{susobasemap}
\alias{susobasemap-package}
\title{susobasemap: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: MIchael Wild \email{Wild@worldbank.org}

}
\keyword{internal}
